+++ env.d.ts
@@ -1,11 +1,13 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
+  readonly VITE_MINK_URL: string;
   readonly VITE_BACKEND_URL: string;
   readonly VITE_AUTH_URL: string;
   readonly VITE_AUTH_GUI_URL: string;
   readonly VITE_LOGOUT_URL: string;
   readonly VITE_METADATA_URL?: string;
+  readonly VITE_KARPS_URL: string;
   readonly VITE_KORP_URL: string;
   readonly VITE_STRIX_URL: string;
   readonly VITE_NEWS_URL: string;