+++ src/components/FileUpload.vue
@@ -15,6 +15,7 @@ const props = defineProps<{
   primary?: boolean;
   accept?: string[];
   multiple?: boolean;
+  replace?: boolean;
   showProgress?: boolean;
 }>();