+++ src/file.ts
@@ -8,13 +8,14 @@ export const FILE_FORMATS = [
   "mp3",
   "ogg",
   "wav",
+  "jsonl",
 ] as const;
 
 export type FileFormat = (typeof FILE_FORMATS)[number];
 
 /** Whether file format is human-readable (i.e. can be displayed in the UI) */
 export function isReadable(format?: string): boolean {
-  return ["txt", "xml", "conllu"].includes(format || "");
+  return ["txt", "xml", "conllu", "jsonl"].includes(format || "");
 }
 
 export const CORPUS_SOURCE_FORMATS = [
@@ -28,8 +29,10 @@ export const CORPUS_SOURCE_FORMATS = [
   "ogg",
   "wav",
 ] as const;
+export const LEXICON_SOURCE_FORMATS = ["jsonl"] as const;
 
 export const SOURCE_FORMATS = {
   corpus: CORPUS_SOURCE_FORMATS,
+  lexicon: LEXICON_SOURCE_FORMATS,
   metadata: [],
 } as const;