+++ src/lexicon/LexiconView.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import useResourceIdParam from "@/resource/resourceIdParam.composable";
+import ResourceViewTemplate from "@/resource/ResourceViewTemplate.vue";
+
+const id = useResourceIdParam();
+</script>
+
+<template>
+  <ResourceViewTemplate type="lexicon" :id />
+</template>